Replace bacon with rspec
